refactor: Muda partes de wavelet transform e thresholding de volta para double para evitar erros de arredondamento

+++ headers/signal_base.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef int sample_t;
+typedef int16_t sample_t;
 typedef sample_t* signal_t;
 
 int sample_compare_function(const void* a, const void* b);
